requirements.txt
pydantic>=2.5.3
SQLAlchemy>=2.0.28
aiogram>=3.4.1
alembic>=1.13.1
black>=24.2.0
cryptography>=42.0.5
emoji>=2.10.1
aiohttp>=3.9.3
python-dateutil>=2.8.2
pytils>=0.4.1
APScheduler>=3.10.4
pytz>=2024.1
cachetools>=5.3.3
pydantic-settings>=2.7.1
asyncmy>=0.2.10
GitPython>=3.1.44